项目/根据 git 变化更新数据库/requirements.txt
GitPython
SQLAlchemy
pymysql